requirements.txt
python-telegram-bot[job-queue]>=20.0
outline-vpn-api>=6.3.0
APScheduler>=3.11.0
python-dotenv>=1.0.0
aiocryptopay>=0.4.7
requests>=2.32.0
aiohttp>=3.12.0
Flask>=2.3.0
gunicorn>=21.0.0
# For YouKassa integration:
yookassa>=3.5.0
# PostgreSQL support:
psycopg2-binary>=2.9.0
# For VPS API communication:
httpx>=0.24.0
flask-cors>=4.0.0